== NOTE == unplayable, only main menu works (a part of it)
These are many changes done over time, simply is:

- new user interface (work in progress): based on layers, scenes and uses different way of sending user events.
- new input handling method (unfinished): input events and actions are separated, will be configurable.
- new font system: uses included freetype, selected font is DroidSans from Android, might need to add the license file. The fonts supports many languages, including CJK characters.
- new UString class.... it is a unicode string class, all text in game should use this class one day...
- new program 'colorgen': generates the tables needed for special effects like blending.
- new program 'fontdatagen': generates the tables needed for the text rendering system.

- in process of removing Lua scripting support, right now only for config.
- in StrManager, removed the tricky UTF conversion, strings shall be unicode.

+++ trunk/netpanzer/src/NetPanzer/Classes/ViewCamera.hpp
@@ -21,6 +21,8 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Interfaces/MapInterface.hpp"
 #include "Interfaces/TileInterface.hpp"
 
+#include "Types/iXY.hpp"
+
 class ViewCamera : private MapInterface
 {
 protected:
@@ -58,9 +60,9 @@ public:
         view_win->setSize(view_size);
     }
 
-    inline long getCameraDistance( const iXY &world_loc )
+    inline unsigned getCameraDistance( const iXY &world_loc )
     {
-        return( (long) (loc - world_loc).mag2() );
+        return (loc - world_loc).mag2();
     }
 };
